First Commit
This folder will take the runtime bits and configuration macros in lan.h
and make them clearer.  Runtime is related to lighting so it will be
included in the lighting .c and .h files.  Other macros that only have
to do with a certain module will be placed in the respective .h files
(like setting up runtime counter on Timer/Counter0, and
enabling/disabling interrupts).  Also need to take the generic names
from the sketch and make them consistent across all files.